Main_App/requirement.txt
Flask==1.1.1
Flask-Bootstrap4==4.0.2
Flask-WTF==0.14.3
pandas==0.25.1
json5==0.8.5
jsonschema==3.0.2
qrcode==6.1
pyodbc==4.0.27